add wrapper function to driver
