+++ experiments/ClimaEarth/user_io/postprocessing.jl
@@ -12,20 +12,18 @@ producing the leaderboard if monthly data is available, performing
 conservation checks if enabled, and closing all diagnostics file writers.
 """
 function postprocess_sim(cs, postprocessing_vars)
-    (; use_coupler_diagnostics, output_default_diagnostics, t_end, conservation_softfail) = postprocessing_vars
+    (; output_default_diagnostics, t_end, conservation_softfail) = postprocessing_vars
     output_dir = cs.dirs.output
     artifact_dir = cs.dirs.artifacts
     coupler_output_dir = joinpath(output_dir, "coupler")
     atmos_output_dir = joinpath(output_dir, "clima_atmos")
     land_output_dir = joinpath(output_dir, "clima_land")
 
-    # Plot generic diagnostics if requested
-    if use_coupler_diagnostics
-        @info "Plotting diagnostics for coupler, atmos, and land"
-        make_diagnostics_plots(coupler_output_dir, artifact_dir, output_prefix = "coupler_")
-        make_diagnostics_plots(atmos_output_dir, artifact_dir, output_prefix = "atmos_")
-        make_diagnostics_plots(land_output_dir, artifact_dir, output_prefix = "land_")
-    end
+    # Plot generic diagnostics
+    @info "Plotting diagnostics for coupler, atmos, and land"
+    make_diagnostics_plots(coupler_output_dir, artifact_dir, output_prefix = "coupler_")
+    make_diagnostics_plots(atmos_output_dir, artifact_dir, output_prefix = "atmos_")
+    make_diagnostics_plots(land_output_dir, artifact_dir, output_prefix = "land_")
 
     # Plot all model states and coupler fields (useful for debugging)
     !CA.is_distributed(cs.comms_ctx) && debug(cs, artifact_dir)
